Version 1.1.1 (#25)
* Version 1.1.1
- Moved Intel workaround to legacy folder, due to hardware error.
- Moved Intel workaround service file too.
- Added blocked dialogue script.
- Added blocked MIME type handlers the XFCE4 way.
- Deleted panel-profile-teams, perhaps in the future - not now.
- Applied the above changes to newly installed NTOS instances through setup-user.sh and setup-root.sh.

* Move docker to docker directory and edit workflows.

* Do not modify custom tagged files.

* Minor changes to code-refresh.sh

* Clarification in error.

* Typo.

* Removed rdp pulling.

* Already inplemented.

* Hotfix for some code.

* symbolic link fix.

* Change script-initiated reboot.

* Update 1.1-1.1.1.sh

* Updating and setting everything ready for 1.1.1

* Should be good for v1.1.1

* Clear more.

* Autoremove add

* Shifting around.

* Trying this

+++ ntos/assets/bin/code-refresh.sh
@@ -1,7 +1,7 @@
 #!/bin/bash
 #
 # Example usage:
-# su root -c "bash /opt/ntos/bin/updater.sh --update --cleanup"
+# su root -c "bash /opt/ntos/bin/code-refresh.sh --update --cleanup"
 #
 args=("$@")
 
@@ -9,17 +9,15 @@ args=("$@")
 source /opt/ntos/VERSION
 #local_version=$VERSION
 #remote_version=$(curl -s "${ORIGIN}"/VERSION | grep "VERSION=" | cut -d "=" -f2-)
-local_rdp=$RDP
 
 relevant_files=(
-    "/opt/ntos/remote-connection.rdp"
-    "/opt/ntos/bin/credcon.sh"
     "/opt/ntos/bin/background-sync.sh"
+    "/opt/ntos/bin/blocked-dialogue.sh"
+    "/opt/ntos/bin/credcon.sh"
     "/opt/ntos/bin/install-firmware.sh"
-    "/opt/ntos/bin/intel-workaround.sh"
-    "/opt/ntos/bin/updater.sh"
-    "/opt/ntos/panel-profile.tar.bz2"
+    "/opt/ntos/bin/code-refresh.sh"
     "/opt/ntos/desktop.png"
+    "/opt/ntos/panel-profile.tar.bz2"
 )
 
 # Function to check cmd arguments.
@@ -36,13 +34,11 @@ contains_arg() {
 pull_latest_code() {
     echo -e '\nDownloading needed files...'
 
-    curl -s "${ORIGIN}"/rdp/"${local_rdp}".rdp > /opt/ntos/remote-connection.rdp.new
-
     curl -s "${ORIGIN}"/credcon/credcon.sh > /opt/ntos/bin/credcon.sh.new
     curl -s "${ORIGIN}"/assets/bin/background-sync.sh > /opt/ntos/bin/background-sync.sh.new
+    curl -s "${ORIGIN}"/assets/bin/blocked-dialogue.sh > /opt/ntos/bin/blocked-dialogue.sh.new
     curl -s "${ORIGIN}"/assets/bin/install-firmware.sh > /opt/ntos/bin/install-firmware.sh.new
-    curl -s "${ORIGIN}"/assets/bin/intel-workaround.sh > /opt/ntos/bin/intel-workaround.sh.new
-    curl -s "${ORIGIN}"/assets/bin/updater.sh > /opt/ntos/bin/updater.sh.new
+    curl -s "${ORIGIN}"/assets/bin/code-refresh.sh > /opt/ntos/bin/code-refresh.sh.new
 
     # Bigger files what are not just text, therefor are downloaded with curl.
     curl -s -o /opt/ntos/panel-profile.tar.bz2.new "${ORIGIN}/assets/panel-profile.tar.bz2"
@@ -55,20 +51,32 @@ upgrade_all() {
         old_file="${file}.old"
 
         if [ -f "$file" ]; then
-            echo "Updating: $file"
-            mv "$file" "$old_file"
-            mv "$new_file" "$file"
-        else
+            if grep -q "^#CUSTOM" "$file"; then
+                echo "Not changing ${file} due to #CUSTOM flag. Current latest version is placed at ${file}.new"
+            else
+                echo "Updating: $file"
+                mv "$file" "$old_file"
+                mv "$new_file" "$file"
+            fi
+        elif [[ -f "$new_file" ]] && [ ! -f "$file" ]; then
             echo "Installing: $file"
             mv -v "$new_file" "$file"
+        else
+            echo "Weird situation, doing nothing... suggest manual intervention."
         fi
     done
 }
 
 cleanup_old() {
     for file in "${relevant_files[@]}"; do
-        echo "Removing remaining: ${file}.old"
-        rm "${file}.old"
+        if [ -f "$file" ]; then
+            echo "Removing remaining: ${file}.old"
+            if [ -f "${file}.old" ]; then
+                rm "${file}.old"
+            fi
+        else
+            echo "Not cleaning up old file: ${file} does not exist."
+        fi
     done
 }
